requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
cryptography>=41.0.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
